paper/requirements.txt
mavsdk==1.4.9
openai>=1.50.0
pydantic>=2.9.0
httpx>=0.27.0
rich>=13.8.0
python-dotenv>=1.0.1
tenacity>=8.5.0
numpy>=1.26.0
